progressbar: add automatic full-width mode
This change required moving get_terminal_size to _termui_impl,
which should not have a too large impact as get_terminal_size
is used inside click as far as I can see (ag'd it) by HelpFormatter.

I added a API wrapper inside termui that imports _termui_impl
only on-demand as with the other functions.
